requirements.txt
python-binance
pandas
openpyxl
websocket
plotly
tqdm
scikit-learn
tensorflow
keras
pymongo
psycopg2; platform_system == "Windows"
psycopg2-binary; platform_system == "Linux"
matplotlib
tqdm
